Change name of crack to cracking. Begin crack.rb CLI
